add tests for DecodeChains
